coreutils: add symlink-usr use flag and mark as stable.

+++ sys-apps/coreutils/coreutils-8.25.ebuild
@@ -23,8 +23,8 @@ SRC_URI="mirror://gnu/${PN}/${P}.tar.xz
 
 LICENSE="GPL-3"
 SLOT="0"
-KEYWORDS="~alpha ~amd64 ~arm ~arm64 ~hppa ~ia64 ~m68k ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc ~x86 ~amd64-fbsd ~x86-fbsd ~arm-linux ~x86-linux"
-IUSE="acl caps gmp hostname kill multicall nls selinux static userland_BSD vanilla xattr"
+KEYWORDS="~alpha amd64 ~arm arm64 ~hppa ~ia64 ~m68k ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc ~x86 ~amd64-fbsd ~x86-fbsd ~arm-linux ~x86-linux"
+IUSE="acl caps gmp hostname kill multicall nls selinux static symlink-usr userland_BSD vanilla xattr"
 
 LIB_DEPEND="acl? ( sys-apps/acl[static-libs] )
 	caps? ( sys-libs/libcap )
@@ -140,7 +140,7 @@ src_install() {
 	insinto /etc
 	newins src/dircolors.hin DIR_COLORS
 
-	if [[ ${USERLAND} == "GNU" ]] ; then
+	if [[ ${USERLAND} == "GNU" ]] && ! use symlink-usr; then
 		cd "${ED}"/usr/bin
 		dodir /bin
 		# move critical binaries into /bin (required by FHS)
@@ -159,7 +159,7 @@ src_install() {
 		for x in ${com} uname ; do
 			dosym /bin/${x} /usr/bin/${x}
 		done
-	else
+	elif [[ ${USERLAND} != "GNU" ]] ; then
 		# For now, drop the man pages, collides with the ones of the system.
 		rm -rf "${ED}"/usr/share/man
 	fi
